@@ -14,16 +14,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.2",
-    "@angular/common": "~12.2.2",
-    "@angular/compiler": "~12.2.2",
-    "@angular/core": "~12.2.2",
-    "@angular/forms": "~12.2.2",
-    "@angular/platform-browser": "~12.2.2",
-    "@angular/platform-browser-dynamic": "~12.2.2",
-    "@angular/platform-server": "~12.2.2",
-    "@angular/router": "~12.2.2",
-    "bootstrap": "^5.1.0",
+    "@angular/animations": "~13.1.2",
+    "@angular/cdk": "^13.1.2",
+    "@angular/common": "~13.1.2",
+    "@angular/compiler": "~13.1.2",
+    "@angular/core": "~13.1.2",
+    "@angular/forms": "~13.1.2",
+    "@angular/material": "^13.1.2",
+    "@angular/platform-browser": "~13.1.2",
+    "@angular/platform-browser-dynamic": "~13.1.2",
+    "@angular/platform-server": "~13.1.2",
+    "@angular/router": "~13.1.2",
+    "@types/wicg-file-system-access": "^2020.9.4",
+    "bootstrap": "^5.1.3",
+    "idb": "^7.0.0",
     "jquery": "^3.5.1",
     "oidc-client": "^1.11.5",
     "popper.js": "^1.16.0",
@@ -33,19 +37,19 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.2",
-    "@angular/cli": "~12.2.2",
-    "@angular/compiler-cli": "~12.2.2",
+    "@angular-devkit/build-angular": "~13.1.3",
+    "@angular/cli": "~13.1.3",
+    "@angular/compiler-cli": "~13.1.2",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.8",
-    "@types/node": "^12.11.1",
+    "@types/node": "^12.20.41",
     "jasmine-core": "~3.8.0",
-    "karma": "~6.3.0",
+    "karma": "^6.3.11",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "typescript": "~4.2.3"
+    "typescript": "~4.5.4"
   },
   "optionalDependencies": {}
 }
