@@ -15,7 +15,7 @@ namespace ED_Virtual_Wing.Data
 
         protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
         {
-            optionsBuilder.UseMySql(Environment.GetEnvironmentVariable("EDDAPI_DB_CONNECTIONSTRING") ?? "server=localhost;user=root;password=1234;database=ef",
+            optionsBuilder.UseMySql(Environment.GetEnvironmentVariable("EDVW_MARIADB_CONNECTIONSTRING") ?? "server=localhost;user=root;password=1234;database=ef",
                 new MariaDbServerVersion(new Version(10, 3, 25)),
                 options =>
                 {
